+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Number fields that are emptied crashed the application
 - Provided more clear 'icon' for the copy layer action
 
+### Added
+
+- Ability to move from frame to frame in the animation screen
+
 ## [1.0.0] - 2021-04-18
 
 On the 18th of January 2021 I started with a project to allow myself to create animated scenery and characters using drawn images. In the past 3 months I learned a lot about WebGL, electron and drawing for animations. I'm really happy with the minimal viable product I have now, so today I'm making the release official.